Allow for bulk ingest from a regular file directory.
