+++ deposit/src/main/java/edu/unc/lib/deposit/work/AbstractDepositJob.java
@@ -362,4 +362,5 @@ public abstract class AbstractDepositJob implements Runnable {
 	public File getSubdir(String subpath) {
 		return new File(getDepositDirectory(), subpath);
 	}
+	
 }
