+++ metadata/src/main/java/edu/unc/lib/dl/util/PackagingType.java
@@ -27,7 +27,8 @@ public enum PackagingType {
 	ATOM("http://purl.org/net/sword/terms/Atom"),
 	BAG_WITH_N3("http://cdr.unc.edu/BAGIT/profiles/N3"),
 	BAGIT("http://purl.org/net/sword/package/BagIt"),
-	PROQUEST_ETD("http://proquest.com");
+	PROQUEST_ETD("http://proquest.com"),
+	DIRECTORY("http://cdr.unc.edu/DirectoryIngest");
 	
 	private String uri;
